Split compare.js into smaller component files
