+++ app/assets/scripts/views/compare/index.js
@@ -1,0 +1,3 @@
+import Compare from './compare';
+
+export default Compare;
